refactor(line): sections를 일급 컬렉션으로 분리
