+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -1,21 +1,13 @@
 package nextstep.subway.line.domain;
 
-import static java.util.Collections.*;
-
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Objects;
-
-import javax.persistence.CascadeType;
 import javax.persistence.Column;
+import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.OneToMany;
 
 import nextstep.subway.common.BaseEntity;
-import nextstep.subway.line.exception.AlreadyInitializedException;
 
 @Entity
 public class Line extends BaseEntity {
@@ -26,8 +18,8 @@ public class Line extends BaseEntity {
     private String name;
     private String color;
 
-    @OneToMany(mappedBy = "line", cascade = CascadeType.ALL)
-    private List<Section> sections = new LinkedList<>();
+    @Embedded
+    private Sections sections = new Sections();
 
     public Line() {
     }
@@ -47,24 +39,10 @@ public class Line extends BaseEntity {
     }
 
     public void addSection(Section section) {
-        validateSection(section);
-        sections.add(section);
+        sections.addSection(section);
         section.setLine(this);
     }
 
-    private void validateSection(Section section) {
-        if (Objects.isNull(section)) {
-            throw new IllegalArgumentException("null 인 구간을 추가할 수는 없습니다.");
-        }
-
-        if (sections.contains(section)) {
-            throw new AlreadyInitializedException("이미 본 라인에 존재하는 구간입니다.");
-        }
-
-        if (section.isAddedToLine()) {
-            throw new AlreadyInitializedException("이미 특정 라인에 포함되어있는 구간입니다.");
-        }
-    }
     // public List<Station> getStations() {
     //     List<Station> stations = new ArrayList<>();
     //     Station station = getStartStation();
@@ -79,10 +57,6 @@ public class Line extends BaseEntity {
     //     return unmodifiableList(stations);
 
     // }
-
-    public List<Section> getSections() {
-        return unmodifiableList(sections);
-    }
 
     public Long getId() {
         return id;
